+++ data-migrations/local-only/2025-11-26-1255-fix-null-user-createdById.sql
@@ -1,0 +1,2 @@
+
+UPDATE users SET createdById = id, modifiedById = id WHERE createdById IS NULL;