+++ src/models/__tests__/Repository.spec.ts
@@ -183,26 +183,6 @@ describe('findBy Queries', () => {
     expect(result).toEqual([]);
   });
 
-  it('findByProjectOutputId should call query with correct params and return the objects', async () => {
-    localQuery.mockResolvedValueOnce([repo]);
-    const id = casual.integer(1, 99);
-    const result = await Repository.findByProjectOutputId('testing', context, id);
-    const sql = 'SELECT r.* FROM repositories r';
-    const joinClause = 'INNER JOIN projectOutputRepositories por ON r.id = por.repositoryId';
-    const whereClause = 'WHERE por.projectOutputId = ?';
-    const vals = [id.toString()];
-    expect(localQuery).toHaveBeenCalledTimes(1);
-    expect(localQuery).toHaveBeenLastCalledWith(context, `${sql} ${joinClause} ${whereClause}`, vals, 'testing')
-    expect(result).toEqual([repo]);
-  });
-
-  it('findByProjectOutputId should return an empty array if there are no records', async () => {
-    localQuery.mockResolvedValueOnce([]);
-    const id = casual.integer(1, 99);
-    const result = await Repository.findByProjectOutputId('testing', context, id);
-    expect(result).toEqual([]);
-  });
-
   it('search should work when a Research Domain, search term and repositoryType are specified', async () => {
     localPaginationQuery.mockResolvedValueOnce([repo]);
     const term = casual.words(3);
@@ -513,87 +493,5 @@ describe('delete', () => {
     expect(Object.keys(result.errors).length).toBe(0);
     expect(result.errors).toEqual({});
     expect(result).toBeInstanceOf(Repository);
-  });
-});
-
-describe('addToProjectOutput', () => {
-  let context;
-  let mockRepository;
-
-  beforeEach(async () => {
-    jest.resetAllMocks();
-
-    context = await buildMockContextWithToken(logger);
-
-    mockRepository = new Repository({
-      id: casual.integer(1, 99),
-      name: casual.words(3),
-      url: casual.url
-    });
-  });
-
-  afterEach(() => {
-    jest.clearAllMocks();
-  });
-
-  it('associates the Repository to the specified ProjectOutput', async () => {
-    const outputId = casual.integer(1, 999);
-    const querySpy = jest.spyOn(Repository, 'query').mockResolvedValueOnce(mockRepository);
-    const result = await mockRepository.addToProjectOutput(context, outputId);
-    expect(querySpy).toHaveBeenCalledTimes(1);
-    let expectedSql = 'INSERT INTO projectOutputRepositories (repositoryId, projectOutputId, createdById,';
-    expectedSql += 'modifiedById) VALUES (?, ?, ?, ?)';
-    const userId = context.token.id.toString();
-    const vals = [mockRepository.id.toString(), outputId.toString(), userId, userId]
-    expect(querySpy).toHaveBeenLastCalledWith(context, expectedSql, vals, 'Repository.addToProjectOutput')
-    expect(result).toBe(true);
-  });
-
-  it('returns null if the domain cannot be associated with the ProjectOutput', async () => {
-    const outputId = casual.integer(1, 999);
-    const querySpy = jest.spyOn(Repository, 'query').mockResolvedValueOnce(null);
-    const result = await mockRepository.addToProjectOutput(context, outputId);
-    expect(querySpy).toHaveBeenCalledTimes(1);
-    expect(result).toBe(false);
-  });
-});
-
-describe('removeFromProjectOutput', () => {
-  let context;
-  let mockRepository;
-
-  beforeEach(async () => {
-    jest.resetAllMocks();
-
-    context = await buildMockContextWithToken(logger);
-
-    mockRepository = new Repository({
-      id: casual.integer(1, 99),
-      name: casual.word,
-      url: casual.url
-    });
-  });
-
-  afterEach(() => {
-    jest.clearAllMocks();
-  });
-
-  it('removes the Repository association with the specified ProjectOutput', async () => {
-    const outputId = casual.integer(1, 999);
-    const querySpy = jest.spyOn(Repository, 'query').mockResolvedValueOnce(mockRepository);
-    const result = await mockRepository.removeFromProjectOutput(context, outputId);
-    expect(querySpy).toHaveBeenCalledTimes(1);
-    const expectedSql = 'DELETE FROM projectOutputRepositories WHERE repositoryId = ? AND projectOutputId = ?';
-    const vals = [mockRepository.id.toString(), outputId.toString()]
-    expect(querySpy).toHaveBeenLastCalledWith(context, expectedSql, vals, 'Repository.removeFromProjectOutput')
-    expect(result).toBe(true);
-  });
-
-  it('returns null if the domain cannot be removed from the ProjectOutput', async () => {
-    const repositoryId = casual.integer(1, 999);
-    const querySpy = jest.spyOn(Repository, 'query').mockResolvedValueOnce(null);
-    const result = await mockRepository.removeFromProjectOutput(context, repositoryId);
-    expect(querySpy).toHaveBeenCalledTimes(1);
-    expect(result).toBe(false);
   });
 });